:ambulance: [Fixes #29] Use Int64 explicitly for Core Data objects
Fixes crashes for 32bit devices like iPhone 5

+++ StudFileProvider/Extensions/FileProvider/FileProviderItemConvertibleState.swift
@@ -23,7 +23,7 @@ public protocol FileProviderItemConvertibleState: class {
     /// Item convertible this state describes.
     var item: FileProviderItemConvertible { get }
 
-    var favoriteRank: Int { get set }
+    var favoriteRank: Int64 { get set }
 
     var lastUsedAt: Date? { get set }
