+++ StudKit/Models/Courses/Course.swift
@@ -59,7 +59,7 @@ public final class Course: NSManagedObject, CDCreatable, CDIdentifiable, CDSorta
     // MARK: Managing Metadata
 
     /// Identifier for this course's group, which determines the course's sorting and color.
-    @NSManaged public var groupId: Int
+    @NSManaged public var groupId: Int64
 
     /// Whether the user chose to hide this course. May be set automatically when hiding a semester.
     @NSManaged public var isHidden: Bool
