+++ StudKit/Models/Courses/CourseState.swift
@@ -38,7 +38,7 @@ public final class CourseState: NSManagedObject, CDCreatable, CDSortable {
 
     // MARK: Managing Metadata
 
-    @NSManaged public var favoriteRank: Int
+    @NSManaged public var favoriteRank: Int64
 
     @NSManaged public var tagData: Data?
 
@@ -47,7 +47,7 @@ public final class CourseState: NSManagedObject, CDCreatable, CDSortable {
     public required convenience init(createIn context: NSManagedObjectContext) {
         self.init(context: context)
 
-        favoriteRank = fileProviderFavoriteRankUnranked
+        favoriteRank = Int64(fileProviderFavoriteRankUnranked)
     }
 
     // MARK: - Sorting
