+++ StudKit/Models/Semesters/SemesterState.swift
@@ -31,7 +31,7 @@ public final class SemesterState: NSManagedObject, CDCreatable, CDSortable {
 
     // MARK: Managing Metadata
 
-    @NSManaged public var favoriteRank: Int
+    @NSManaged public var favoriteRank: Int64
 
     @NSManaged public var tagData: Data?
 
@@ -46,7 +46,7 @@ public final class SemesterState: NSManagedObject, CDCreatable, CDSortable {
     public required convenience init(createIn context: NSManagedObjectContext) {
         self.init(context: context)
 
-        favoriteRank = fileProviderFavoriteRankUnranked
+        favoriteRank = Int64(fileProviderFavoriteRankUnranked)
     }
 
     // MARK: - Sorting
